requirements.txt
## The following requirements were added by pip freeze:
absl-py==1.4.0
attrs==22.2.0
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
cloudpickle==2.2.0
Deprecated==1.2.13
docker==6.0.1
docstring-parser==0.15
fire==0.5.0
google-api-core==2.11.0
google-auth==2.16.0
google-cloud-aiplatform==1.20.0
google-cloud-bigquery==2.34.4
google-cloud-core==2.3.2
google-cloud-resource-manager==1.8.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.58.0
grpc-google-iam-v1==0.12.6
grpcio==1.51.1
grpcio-status==1.48.2
idna==3.4
jsonschema==3.2.0
kfp==2.0.0b10
kfp-pipeline-spec==0.1.17
kfp-server-api==2.0.0a6
kubernetes==23.6.0
oauthlib==3.2.2
packaging==21.3
proto-plus==1.22.2
protobuf==3.20.3
pyasn1==0.5.0rc2
pyasn1-modules==0.3.0rc1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
PyYAML==5.4.1
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
strip-hints==0.1.10
tabulate==0.9.0
termcolor==2.2.0
typer==0.7.0
uritemplate==3.0.1
urllib3==1.26.14
websocket-client==1.4.2
wrapt==1.15.0rc1
